requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.11.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
environs==9.5.0
gunicorn==20.1.0
django-filter~=23.3